feat(lexarg): Re-export lexarg_error
