+++ crates/lexarg-parser/README.md
@@ -1,0 +1,26 @@
+# lexarg
+
+> Minimal, API stable CLI parser"
+
+[![Documentation](https://img.shields.io/badge/docs-master-blue.svg)][Documentation]
+![License](https://img.shields.io/crates/l/lexarg.svg)
+[![Crates Status](https://img.shields.io/crates/v/lexarg.svg)](https://crates.io/crates/lexarg)
+
+## License
+
+Licensed under either of
+
+* Apache License, Version 2.0, ([LICENSE-APACHE](LICENSE-APACHE) or <http://www.apache.org/licenses/LICENSE-2.0>)
+* MIT license ([LICENSE-MIT](LICENSE-MIT) or <http://opensource.org/licenses/MIT>)
+
+at your option.
+
+### Contribution
+
+Unless you explicitly state otherwise, any contribution intentionally
+submitted for inclusion in the work by you, as defined in the Apache-2.0
+license, shall be dual-licensed as above, without any additional terms or
+conditions.
+
+[Crates.io]: https://crates.io/crates/lexarg
+[Documentation]: https://docs.rs/lexarg
