+++ crates/lexarg/Cargo.toml
@@ -27,6 +27,7 @@ pre-release-replacements = [
 default = []
 
 [dependencies]
+lexarg-parser = { "version" = "0.1.0", path = "../lexarg-parser" }
 
 [dev-dependencies]
